+++ src/js/index.js
@@ -1,3 +1,5 @@
+import "./recovery"
+
 import "Resources/styles/fonts.css"
 import "Resources/styles/constructor.css"
 
@@ -5,5 +7,6 @@ import "./loaders/MainLoader"
 
 import App from "./main/app"
 import Report from "./main/report"
+
 
 Report.write(`App ${App.fullName}: Build date ${App.buildDate}`)