+++ src/js/services/Settings/user/manager.js
@@ -1,4 +1,5 @@
 import SettingsLayout from "./layout"
+import Report from "../../../main/report"
 
 
 export default class SettingsLayoutManager {
@@ -8,6 +9,7 @@ export default class SettingsLayoutManager {
         if (!(l instanceof SettingsLayout)) throw new TypeError("Only Settings Layout can be applied")
 
         this._layout = l
+        Report.write("Settings layout applied")
     }
 
     static get layout() {