+++ src/js/tools/transformation/text/SecondsToTime.js
@@ -1,0 +1,7 @@
+export default function SecondsToTime(time) {
+    const d = Number(time)
+    const h = Math.floor(d / 3600)
+    const m = Math.floor((d % 3600) / 60)
+    const s = Math.floor(d % 3600 % 60)
+    return [d, h, m, s]
+}